requirements.txt
pymongo==4.6.1
twscrape~=0.11.1
beautifulsoup4==4.12.3
selenium==4.16.0
requests==2.31.0
python-dotenv==1.0.0
click==8.1.3
prometheus_client~=0.20.0
pycountry~=22.1.10